+++ src/App.svelte
@@ -10,7 +10,6 @@
   import PotentialMatches from "./components/PotentialMatches.svelte";
   import Matches from "./components/Matches.svelte";
   import Messages from "./components/Messages.svelte";
-  import MyListings from "./components/MyListings.svelte";
 </script>
 
 <Router>
@@ -44,10 +43,6 @@
 
   <PrivateRoute path="new-listing" let:location>
     <NewListing />
-  </PrivateRoute>
-
-  <PrivateRoute path="my-listings" let:location>
-    <MyListing />
   </PrivateRoute>
 
   <PrivateRoute path="messages" let:location>
